Factor

A wrapper for factor with enhanced control.

cfactor(x, levels, labels = levels, exclude = NA, ordered = is.ordered(x),
  nmax = NA, sep = c("-", "to"))

Arguments

x

a vector of data, usually taking a small number of distinct values.

levels

an optional vector of the values (as character strings) that x might have taken. If missing (the default), the levels are determined by sorting the numerical elements in x. This is not as in factor. See 'Details'.

labels

either an optional character vector of labels for the levels (in the same order as levels after removing those in exclude), or a character string of length 1.

exclude

a vector of values to be excluded when forming the set of levels. This should be of the same type as x, and will be coerced if necessary.

ordered

logical flag to determine if the levels should be regarded as ordered (in the order given).

nmax

an upper bound on the number of levels; see ‘Details’.

sep

A character vector giving all strings that are used to separate the lower numerical boundary of a range from the upper numerical boundary within x. See 'Details' for more information.

Value

See factor.

Details

cfactor wraps factor but provides enhanced control. The enhanced control has the following elements:

more warnings: NAs are never produced silently. Empty levels also produce a warning. If levels and labels intersect, a warning is produced too.

flexible order detection: By default, the order of the levels is given by the order of the numerical elements in x. See below.

Levels of the factor or determined by the levels argument. If levels is missing, the order of the levels is determined as follows:

If sep is set to NULL: Levels are determined by sorting the unique values in x, which is the same behaviour as in factor.

If sep is set to NA, the order is given by first appearance in x.

Otherwise,

If every value in x contains numbers: Levels are determined by sorting the numerical values preceeding the separators indicated in sep.

If not every value in x contains numbers: Levels are determined by sorting the unique values in x, which is the same behaviour as in factor and corresponds to setting sep to NULL.

Apart from the newly introduced argument sep, cfactor has the same arguments and defaults as factor.

Where levels is not supplied, unique is called. Since factors typically have quite a small number of levels, for large vectors x it is helpful to supply nmax as an upper bound on the number of unique values.

Examples

## Not run: ------------------------------------
# ## create warnings with unused and removed levels
# string <- c("a", "b", "c")
# cfactor(string, levels = c("b", "c", "d"))
## ---------------------------------------------

## detecting levels: compare factor and cfactor
hard_to_detect <- c("EUR 21 - EUR 22", "EUR 100 - 101",
"EUR 1 - EUR 10", "EUR 11 - EUR 20")
factor(hard_to_detect, ordered = TRUE)

#> [1] EUR 21 - EUR 22 EUR 100 - 101   EUR 1 - EUR 10  EUR 11 - EUR 20
#> 4 Levels: EUR 1 - EUR 10 < EUR 100 - 101 < ... < EUR 21 - EUR 22

cfactor(hard_to_detect, ordered = TRUE)

#> [1] EUR 21 - EUR 22 EUR 100 - 101   EUR 1 - EUR 10  EUR 11 - EUR 20
#> 4 Levels: EUR 1 - EUR 10 < EUR 11 - EUR 20 < ... < EUR 100 - 101