Decode numerical into categorical data

Decode numerical columns in a data frame into (ordered) factors given the encoding in another data frame.

index_cfactor(data, index, variable = "variable", encoding = "encoding",
  label = "label", ...)

Arguments

data

A data frame containing at least one integer column to decode.

index

A data frame containing the names of the variable to encode, the encoding for data and labels to assign.

variable

The name of the column in index that indicates the variable.

encoding

The name of the column in index that indicates the encoding.

label

The name of the column in index that indicates the label that will be given.

...

Further arguments to be passed to or from other methods, in particular to cfactor.

Value

The original data frame is returned whereas the variables for which an encoding was provided are turned into (ordered) factors. All other columns are returned unmodified.

Details

Arguments passed via ... to cfactor are only recycled if of length 1. Otherwise, an error is thrown. All arguments passed via ... are applied in the order of the data columns but columns not to convert are skipped (see example).

Examples

data <- data.frame(var1 = sample(x = 1:10, size = 20, replace = TRUE),
                   var2 = rep(1:2, 20),
                   var3 = sample(20),
                   var4 = 2,
                   var5 = sample(row.names(USArrests), size = 20),
                   stringsAsFactors = FALSE)

 index <- data.frame(var = rep(paste0("var", 1:3), c(10, 2, 20)),
                   encoding = c(1:10, 1:2, 1:20),
                   label = c(letters[1:10], c("male", "female"), LETTERS[1:20]))

 index_cfactor(data = data, index = index, variable = "var", ordered = c(TRUE, TRUE, FALSE))

#> Warning: the following levels were empty: 
#>  7

#>    var1   var2 var3 var4          var5
#> 1     h   male    D    2    Washington
#> 2     h female    S    2      Maryland
#> 3     j   male    J    2        Hawaii
#> 4     j female    R    2      New York
#> 5     d   male    E    2       Arizona
#> 6     e female    G    2       Vermont
#> 7     d   male    F    2   Connecticut
#> 8     b female    A    2     Minnesota
#> 9     f   male    N    2      Oklahoma
#> 10    e female    P    2    New Jersey
#> 11    h   male    M    2       Georgia
#> 12    c female    T    2      Michigan
#> 13    h   male    K    2       Wyoming
#> 14    a female    L    2        Alaska
#> 15    d   male    C    2      Kentucky
#> 16    i female    O    2 New Hampshire
#> 17    c   male    H    2        Nevada
#> 18    f female    Q    2       Indiana
#> 19    d   male    I    2          Utah
#> 20    f female    B    2      Illinois
#> 21    h   male    D    2    Washington
#> 22    h female    S    2      Maryland
#> 23    j   male    J    2        Hawaii
#> 24    j female    R    2      New York
#> 25    d   male    E    2       Arizona
#> 26    e female    G    2       Vermont
#> 27    d   male    F    2   Connecticut
#> 28    b female    A    2     Minnesota
#> 29    f   male    N    2      Oklahoma
#> 30    e female    P    2    New Jersey
#> 31    h   male    M    2       Georgia
#> 32    c female    T    2      Michigan
#> 33    h   male    K    2       Wyoming
#> 34    a female    L    2        Alaska
#> 35    d   male    C    2      Kentucky
#> 36    i female    O    2 New Hampshire
#> 37    c   male    H    2        Nevada
#> 38    f female    Q    2       Indiana
#> 39    d   male    I    2          Utah
#> 40    f female    B    2      Illinois